ASI-289 WIP code refactor - add ArchivesSpaceToFolio module, move the logic for enhancing marc records from FOLIO client to FolioSynchronizer, pass enhanced marc to FOLIO method
