+++ lib/folio_sync/folio/client.rb
@@ -24,31 +24,8 @@ class FolioSync::Folio::Client < FolioApiClient
     self.find_marc_record(instance_record_hrid: hrid)
   end
 
-  def create_or_update_folio_record(bibid)
-    folio_marc = get_marc_record(bibid)
-
-    if folio_marc
-      update_existing_folio_marc_record(bibid, folio_marc)
-    else
-      create_new_folio_marc_record(bibid)
-    end
-  end
-
-  # If the record doesn't exist in FOLIO, create a new one
-  def create_new_folio_marc_record(bibid)
-    # TODO: Should also update 035 field
-    marc_record = FolioSync::Folio::MarcRecord.new(bibid)
-    marc_record.process_record
-
-    # TODO: Call the FOLIO API to create a new record
-  end
-
-  # If the record already exists in FOLIO, update it
-  def update_existing_folio_marc_record(bibid, folio_marc)
-    marc_record = FolioSync::Folio::MarcRecord.new(bibid, folio_marc)
-    marc_record.process_record
-
-    # TODO: Call the FOLIO API to update the record
+  def create_or_update_folio_record(marc_record)
+    # TODO: Call the FOLIO API to create/update a record
   end
 
   private
