+++ spec/folio_sync/folio/client_spec.rb
@@ -76,94 +76,7 @@ RSpec.describe FolioSync::Folio::Client do
     end
   end
 
-  describe '#create_or_update_folio_record' do
-    before do
-      allow(instance).to receive(:get_marc_record).with(bibid).and_return(nil)
-      allow(instance).to receive(:create_new_folio_marc_record)
-      allow(instance).to receive(:update_existing_folio_marc_record)
-    end
-
-    it 'creates a new FOLIO MARC record if none exists' do
-      instance.create_or_update_folio_record(bibid)
-      expect(instance).to have_received(:create_new_folio_marc_record).with(bibid)
-    end
-
-    it 'updates an existing FOLIO MARC record if one exists' do
-      allow(instance).to receive(:get_marc_record).with(bibid).and_return(marc_record)
-      instance.create_or_update_folio_record(bibid)
-      expect(instance).to have_received(:update_existing_folio_marc_record).with(bibid, marc_record)
-    end
-  end
-
-  describe '#create_new_folio_marc_record' do
-    let(:bibid) { '123456' }
-    let(:marc_file_path) { Rails.root.join('tmp/marc_files', "#{bibid}.xml") }
-    let(:mock_marc_xml) do
-      <<-XML
-        <record xmlns="http://www.loc.gov/MARC21/slim">
-          <controlfield tag="001">123456</controlfield>
-          <datafield tag="245" ind1="1" ind2="0">
-            <subfield code="a">Mock Title</subfield>
-          </datafield>
-        </record>
-      XML
-    end
-
-    # Ensure the tmp/marc_files directory exists
-    # and create a mock MARC file
-    before do
-      FileUtils.mkdir_p(File.dirname(marc_file_path))
-      File.write(marc_file_path, mock_marc_xml)
-
-      allow(FolioSync::Folio::MarcRecord).to receive(:new).and_call_original
-    end
-
-    # Clean up the mock MARC file
-    after do
-      File.delete(marc_file_path) if File.exist?(marc_file_path)
-    end
-
-    it 'creates a new MARC record' do
-      expect {
-        instance.create_new_folio_marc_record(bibid)
-      }.not_to raise_error
-      expect(FolioSync::Folio::MarcRecord).to have_received(:new).with(bibid)
-    end
-  end
-
-  describe '#update_existing_folio_marc_record' do
-    let(:bibid) { '123456' }
-    let(:marc_file_path) { Rails.root.join('tmp/marc_files', "#{bibid}.xml") }
-    let(:mock_marc_xml) do
-      <<-XML
-        <record xmlns="http://www.loc.gov/MARC21/slim">
-          <controlfield tag="001">123456</controlfield>
-          <datafield tag="245" ind1="1" ind2="0">
-            <subfield code="a">Mock Title</subfield>
-          </datafield>
-        </record>
-      XML
-    end
-    let(:folio_marc) { double('FolioMarc') }
-
-    # Ensure the tmp/marc_files directory exists
-    # and create a mock MARC file
-    before do
-      FileUtils.mkdir_p(File.dirname(marc_file_path))
-      File.write(marc_file_path, mock_marc_xml)
-      allow(FolioSync::Folio::MarcRecord).to receive(:new).and_call_original
-    end
-
-    # Clean up the mock MARC file
-    after do
-      File.delete(marc_file_path) if File.exist?(marc_file_path)
-    end
-
-    it 'updates an existing MARC record' do
-      expect {
-        instance.update_existing_folio_marc_record(bibid, folio_marc)
-      }.not_to raise_error
-      expect(FolioSync::Folio::MarcRecord).to have_received(:new).with(bibid, folio_marc)
-    end
-  end
+  # This method has not been implemented yet
+  # describe '#create_or_update_folio_record' do
+  # end
 end
